+++ api/templates/embeds/volunteer_embed.html
@@ -112,11 +112,12 @@
     async function postVolunteer() {
         let data = new FormData(document.getElementById("taeb_server_volunteer_form_embed"));
         
-        const response = await fetch(`https://taeb-server.vercel.app/volunteer`, {
+        const response = await fetch(`http://127.0.0.1:5000/volunteer`, {
             method: "POST",
             body: data
         });
         const result = await response.json();
+        console.log(result);
         
         return result;
     }
