+++ api/volunteer.py
@@ -1,9 +1,11 @@
-from flask import Blueprint, request, jsonify
+from flask import Blueprint, jsonify
+import resend.exceptions
 from webargs.flaskparser import abort, parser, use_args, use_kwargs
-from .data import state_dict, activities, trim_inputs, no_duplicates
+from .data import state_dict, activities, trim_inputs, no_duplicates, rendered_email
 from marshmallow import Schema, fields, validate, pre_load, ValidationError
 from .db import supabase_admin
 from supabase import PostgrestAPIError
+import resend
 
 bp = Blueprint('volunteer', __name__, url_prefix='/volunteer')
 
@@ -34,14 +36,40 @@ class VolunteerSchema(Schema):
 @bp.route('/', methods=["POST"])
 @use_args(VolunteerSchema(), location='form')
 def add_volunteer(args):
+    # add volunteer to database
     response = (
         supabase_admin.table("volunteers")
         .insert(args)
         .execute()
     )
     
+    # get contents of email from database and add in user's name
+    content_response = (
+        supabase_admin.table("content")
+        .select("content")
+        .eq("name", "volunteer_form")
+        .execute()
+    )
+    email_content = content_response.data[0]["content"]
+    email_content = email_content.replace("_name_", f"{args["first_name"]} {args["last_name"]}")
+
+    # Send email to user
+    params: resend.Emails.SendParams = {
+        "from": "Freshta Taeb <onboarding@taebforassembly.com>",
+        "to": [args["email"]],
+        "subject": "Volunteer Onboarding",
+        "html": rendered_email(email_content)
+    }
+    email_success = True
+    try:
+        resend.Emails.send(params)
+    except resend.exceptions.ResendError:
+        email_success = False
+
+    # send back response
     return {
         "success" : True,
+        "email_success": email_success,
         "data" : response.data
     }
